3.2-crud/stocks_products/requirements.txt
django-filter
django
djangorestframework
psycopg2-binary
django-environ
gunicorn
